feat: create graphql-yoga server with tests
