+++ packages/graphql-api-server/package.json
@@ -19,7 +19,6 @@
   ],
   "scripts": {
     "build": "tsup",
-    "schema": "ttoss-graphql-api build-schema",
     "test": "jest"
   },
   "sideEffects": false,
